+++ web/client/translations/data.de-DE.json
@@ -2847,6 +2847,10 @@
             "description": "Ermöglicht das Teilen der Karte auf verschiedene Arten (Link, QR-Code, Einbettung, soziale Netzwerke ...)",
             "title": "Freigabe-Tool"
           },
+          "StreetView": {
+            "title": "Straßenansicht",
+            "description": "Tool zum Durchsuchen von Google Street View-Bildern auf der Karte"
+          },
           "StyleEditor": {
             "description": "Stileditor der Ebene",
             "title": "Stil-Editor"
@@ -3405,6 +3409,17 @@
           "successRemoved": "Benutzersitzung entfernt",
           "remove": "Benutzersitzung zurücksetzen",
           "confirmRemove": "Möchten Sie die Standardkonfiguration wirklich wiederherstellen?"
+        },
+        "streetView": {
+          "title": "Straßenansicht",
+          "description": "Tool zum Durchsuchen von Google Street View-Bildern auf der Karte",
+          "tooltip": "Street-View-Tool öffnen",
+          "emptyTitle": "Straßenansicht",
+          "emptyDescription": "Klicken Sie auf die Karte, um die Bilder der Straßenansicht zu durchsuchen",
+          "messages": {
+            "noDataForPosition": "keine Street-View-Daten für diese Position",
+            "unknownError": "Unbekannter Fehler, siehe Konsole"
+          }
         }
     }
 }
